rebase on apache, fixes some issues in MockRestoreConsumer
