+++ streams/src/test/java/org/apache/kafka/test/ProcessorTopologyTestDriver.java
@@ -284,23 +284,8 @@ public class ProcessorTopologyTestDriver {
      * @return the mock consumer; never null
      */
     protected MockConsumer<byte[], byte[]> createRestoreConsumer(int id, String... storeNames) {
-        MockConsumer<byte[], byte[]> consumer = new MockConsumer<byte[], byte[]>(OffsetResetStrategy.LATEST) {
-            @Override
-            public synchronized void seekToEnd(TopicPartition... partitions) {
-                // do nothing ...
-            }
+        MockConsumer<byte[], byte[]> consumer = new MockRestoreConsumer();
 
-            @Override
-            public synchronized void seekToBeginning(TopicPartition... partitions) {
-                // do nothing ...
-            }
-
-            @Override
-            public synchronized long position(TopicPartition partition) {
-                // do nothing ...
-                return 0L;
-            }
-        };
         // For each store name ...
         for (String storeName : storeNames) {
             String topicName = storeName;
